Change User Model Structure and other Schemas

+++ api/Controllers/UserController.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 // for database context
@@ -10,6 +11,7 @@ using api.Helper;
 using api.Interfaces;
 using api.Mappers;
 using api.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -29,6 +31,8 @@ namespace api.Controllers
             //_context = context; // initialize the database context through dependency injection
         }
 
+
+        [Authorize]
         [HttpGet]
         public async Task<IActionResult> GetAllUsers(QueryObject query)
         {
@@ -41,12 +45,19 @@ namespace api.Controllers
             return Ok(usersInView); //200 OK status code with the list of users in the response body
         }
 
+        [Authorize]
         [HttpGet]
-        [Route("{userId:guid}")] // api/User/{userId}
-        public async Task<IActionResult> GetUserById([FromRoute] Guid userId)
+        [Route("me")] // api/User/{userId}
+        public async Task<IActionResult> GetUserById()
         {
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
+
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (userId == null)
+            {
+                return Unauthorized();
+            }
 
             var user = await _userRepo.GetUserByIdAsync(userId); // FirstOrDefault also works, but Find is more efficient because it uses the primary key to find the user.
             if (user == null)
@@ -56,26 +67,19 @@ namespace api.Controllers
             return Ok(user.ToUserInView()); //200 OK status code with the user in the response body
         }
 
-        // [HttpPost]
-        // //FromBody is used to bind the request body to the user object
-        // //Taking request body object and mapping it to the User model using the mapper class, then saving it to the database.
-        // public async Task<IActionResult> CreateUser([FromBody] UserInCreate userInCreateObject)
-        // {
-        //     if (!ModelState.IsValid)
-        //         return BadRequest(ModelState);
-
-        //     var user = await _userRepo.CreateUserAsync(userInCreateObject); //map the request body to the User model using the mapper class
-        //     return CreatedAtAction(nameof(GetUserById), new { userId = user.UserId }, user.ToUserInView());
-        //     //201 Created status code with the user in the response body
-        // }
-
-
+        [Authorize]
         [HttpPut]
-        [Route("{userId:guid}")]
-        public async Task<IActionResult> UpdateUser([FromRoute] Guid userId, [FromBody] UserInUpdate userInUpdateObject)
+        [Route("me")]
+        public async Task<IActionResult> UpdateUser([FromBody] UserInUpdate userInUpdateObject)
         {
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
+
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (userId == null)
+            {
+                return Unauthorized();
+            }
 
             var userUpdates = await _userRepo.UpdateUserAsync(userId, userInUpdateObject);
             if (userUpdates == null)
@@ -85,13 +89,19 @@ namespace api.Controllers
             return Ok(userUpdates.ToUserInView());
         }
 
-
+        [Authorize]
         [HttpDelete]
-        [Route("{userId:guid}")]
-        public async Task<IActionResult> DeleteUser([FromRoute] Guid userId)
+        [Route("me")]
+        public async Task<IActionResult> DeleteUser()
         {
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
+
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (userId == null)
+            {
+                return Unauthorized();
+            }
 
             var user = await _userRepo.DeleteUserAsync(userId);
             if (user == null)
